% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{balance_f_test}
\alias{balance_f_test}
\title{Test for balance across treatment arms}
\usage{
balance_f_test(data, control, trt, covars)
}
\arguments{
\item{data}{(data frame) Data frame that contains survey data}

\item{control}{(character) Name of control group}

\item{trt}{(character) Name of treatment group}

\item{covars}{(character vector) Names of covariates in survey dataset that
will be used for the covariate balance test.}
}
\description{
Fit OLS regression (one per treatment arm, with a sample that includes
the control and that treatment arm) of the treatment indicators on a set
of covariates (dummy variables for blocks will automatically be included)
and calculate a heteroskedasticity-robust Wald statistic for the hypothesis
that all the coefficients on the covariates (excluding block dummies) are zero.
}
