% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{balance_table}
\alias{balance_table}
\title{Create balance tables}
\usage{
balance_table(data, control, trt, covars, output_file)
}
\arguments{
\item{data}{(data frame) Data frame that contains survey data}

\item{control}{(character) Name of control group}

\item{trt}{(character) Name of treatment group}

\item{covars}{(character vector) Names of covariates in survey dataset that
will be used for the covariate balance test.}

\item{output_file}{(character) Name of file where table will be stored (in
latex format)}
}
\description{
This function creates a balance table in which covariate proportions are listed in
rows and columns correspond to the control and treatment arms.
An additional column contains the treatment-control difference and a p-value
for a t-test of the difference in means.
}
