% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{baseline_lasso}
\alias{baseline_lasso}
\title{Lasso selection of variables}
\usage{
baseline_lasso(
  d,
  dv,
  lasso_covars,
  treatment_var,
  trt,
  control,
  select_factor = FALSE,
  seed = 2020
)
}
\arguments{
\item{d}{(data frame) Data frame that contains the data}

\item{dv}{(character) Name of the DV}

\item{lasso_covars}{(character vector) List of covariates to add to lasso
model for selection}

\item{treatment_var}{(character) Name of treatment group assignment variable.
Default "treatment_group"}

\item{trt}{(character) Name of treatment group}

\item{control}{(character) Name of control group}

\item{select_factor}{(logical) If \code{TRUE}, will include the whole factor from LASSO
in output, if \code{FALSE} it will output just the selected levels from the factor}

\item{seed}{(numeric) Seed for random numbers for repeatability.  Default 2020.}
}
\description{
A simple function to select variables for baseline_model using Lasso
}
