% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{baseline_model}
\alias{baseline_model}
\title{Estimate treatment effects using a baseline model with lasso covariate
selection}
\usage{
baseline_model(
  d,
  dv,
  dv_pre = NULL,
  lasso_covars,
  model_covars,
  treatment_var = "treatment_group",
  trt,
  control,
  dv_label = NULL,
  model_type = "all",
  weights_var = NULL,
  iv = NULL,
  select_factor = TRUE,
  convert_to_df = TRUE,
  treated_bin = TRUE,
  factor_to_bin = FALSE,
  impute = TRUE,
  deact_impute = FALSE,
  standardize = TRUE,
  proportion = FALSE,
  digits = 3,
  se_type = "HC2",
  leebounds = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{d}{(data frame) Data frame that contains the data}

\item{dv}{(character) Name of the DV}

\item{dv_pre}{(character) Name of the pre-treatment DV. \code{NULL} if not
available. Default NULL.}

\item{lasso_covars}{(character vector) List of covariates to add to lasso
model for selection}

\item{model_covars}{(character vector) List of covariates to add to baseline
model, regardless of whether they were selected by the lasso or not.}

\item{treatment_var}{(character) Name of treatment group assignment variable.
Default 'treatment_group'}

\item{trt}{(character) Name of treatment group}

\item{control}{(character) Name of control group}

\item{dv_label}{(character) Label for the DV, which can then be used to create
tables or graphs. Default is same as \code{dv}.}

\item{model_type}{(character) Model type(s) to return.  Must be one of 'SATE',
'LATE', 'PLATE', 'PATE', or 'all' (default). You can also select multiple
models by adding the requested types as a character vector.}

\item{weights_var}{(character) Name of the weights variable in the dataset.
Must be included if \code{model_type} is 'all', 'PATE', or 'PLATE'.}

\item{iv}{(character) Name in the dataset of the variable instrumented by the
treatment. Must be included if \code{model_type} is 'all' or 'LATE'.}

\item{select_factor}{(logical) If \code{TRUE}, will include the whole factor from LASSO
in output, if \code{FALSE} it will output just the selected levels from the factor}

\item{convert_to_df}{(logical) If \code{TRUE}, will check if d is a data frame
and convert d to data frame if check returns FALSE}

\item{treated_bin}{(logical) If \code{TRUE}, will create a binary for the
whether obs belongs to treated group or not}

\item{factor_to_bin}{(logical) If \code{TRUE}, will create binary covariates
from factor (or character) covariates (excluding one category per factor var)}

\item{impute}{(boolean) Whether to use mean imputation for missing values}

\item{deact_impute}{(boolean) Whether to use mean imputation protocol outlined in
deactivation PAP}

\item{standardize}{(logical) If \code{TRUE}, the DV will be standardized
prior to fit the models, and the returning treatment effect estimate will be
a standardized coefficient.}

\item{proportion}{(logical) If \code{TRUE}, will multiply the DV by 100 prior
to fitting the model. Useful if the DV is a percentage; so that the output
is easier to read.}

\item{digits}{(numeric) Number of decimal digits to display in output.}

\item{se_type}{(character) Type of standard errors for treatment effect
estimates. This is inherited from estimatr. Valid types are: "HC0", "HC1",
"HC2", and "HC3". Default is "HC2"}

\item{leebounds}{(boolean) Whether to include Lee bounds for the estimates.}

\item{verbose}{(boolean) Whether to print messages with estimated effects.}
}
\description{
This function follows a two-step procedure. First, it estimates a lasso model
where the DV is regressed on its pre-treatment measure (if available), as well
as a set of covariates measured prior to treatment. Then, it fits a linear
model that includes the treatment indicator, the covariates selected by the
lasso model and additional covariates. The function returns both
weighted (PATE) and unweighted (SATE) treatment effects.
}
