% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passive.R
\name{col_types_passive}
\alias{col_types_passive}
\title{Create column specification for passive monitoring data, to be used
internally by \code{read_csv} to properly parse survey files.}
\usage{
col_types_passive(source)
}
\arguments{
\item{source}{(character) "web" or "app"}
}
\description{
Create column specification for passive monitoring data, to be used
internally by \code{read_csv} to properly parse survey files.
}
