% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Convert data d into dataframe if needed}
\usage{
convert_to_df(d)
}
\arguments{
\item{d}{(data frame) Data frame that contains survey data}
}
\description{
The function checks if d is data frame, if check returns FALSE,
function converts d intro a data frame before returning d
}
