% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{create_composite_scale}
\alias{create_composite_scale}
\title{Create composite scale using factor analysis}
\usage{
create_composite_scale(
  svy,
  items,
  control_group = NULL,
  nfactors = 1,
  impute = FALSE,
  impute_function = "median",
  missingness_report = TRUE,
  missingness_baseline = "study",
  verbose = TRUE
)
}
\arguments{
\item{svy}{(data frame) Data frame that contains survey data}

\item{items}{(character vector) List of variables that will be jointly scaled.}

\item{control_group}{(character) Value of the \code{treatment_group} variable
in the \code{svy} data frame that corresponds to the control group in the
study. If \code{NULL}, the composite scale will be computed using all
observations in the survey dataset.}

\item{nfactors}{(numeric) Number of components to be extracted. This is only
for exploratory purposes; the function will always return a single score.}

\item{impute}{(Boolean) If \code{TRUE}, impute missing values using the
median.}

\item{impute_function}{(character) Either 'median' or 'mean'. Function used
to replace missing values when imputing.}

\item{missingness_report}{(Boolean) If \code{TRUE}, returns report with
details about missing values in each item and across respondents.}

\item{missingness_baseline}{(character) Either 'study' to compute proportion
missing with respect to entire dataset; or 'W2', 'W3', 'W4', 'W5' to compute
it with respect to respondents who completed those waves}

\item{verbose}{(Boolean) If \code{TRUE} (default), print factor loadings.}
}
\description{
Create composite scale from multiple survey items using Principal Component
Analysis with varimax rotation.
}
