% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{descriptive_table}
\alias{descriptive_table}
\title{Create a table with the distribution of a set of covariates from the log data,
comparing control and treatment groups.}
\usage{
descriptive_table(
  data,
  var_labels,
  vars,
  control,
  control_label,
  trt,
  trt_labels,
  digits = 2,
  proportion = FALSE,
  listwise_deletion = TRUE,
  lasso_covars,
  model_covars,
  caption,
  output_file,
  append = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{(data frame) Data frame that contains survey data}

\item{var_labels}{(character vector) Labels for each variable}

\item{vars}{(character vector) Names of covariates in dataset whose
distribution wants to be dislayed in the table}

\item{control}{(character) Name of control group}

\item{control_label}{(character) Label for control group}

\item{trt}{(character vector) Name of treatment group(s)}

\item{trt_labels}{(character vector) Labels for treatment group(s)}

\item{digits}{(numeric) Number of decimal digits to display in table}

\item{proportion}{(logical) Whether output is a percentage or not. If \code{TRUE},
quantities will be multiplied by 100.}

\item{listwise_deletion}{(logical) Whether all metrics should be computed
for the same dataset; which implies that units with a missing value in any
of the variables will be excluded for all variables. If \code{FALSE}, each
metric is computed for the non-missing values of that variable.}

\item{lasso_covars}{(character vector) List of covariates to add to lasso
model for selection in covariate-adjusted p-value.}

\item{model_covars}{(character vector) List of covariates to add to baseline
model, regardless of whether they were selected by the lasso or not, in
order to compute covariate-adjusted p-value.}

\item{caption}{(character) Caption to be added to the latex table}

\item{output_file}{(character) Name of file where table will be stored (in
latex format)}

\item{append}{(logical) If \code{TRUE}, output table will be appended to
existing file. Otherwise, the file will be overwritten.}

\item{verbose}{(logical) If \code{TRUE}, the function will output additional
information, such as the proportion of units with missing values.}
}
\description{
Create a table with the distribution of a set of covariates from the log data,
comparing control and treatment groups.
}
