% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display}
\alias{display}
\title{Display numbers in a pretty format}
\usage{
display(num, pct = FALSE, digits = 3)
}
\arguments{
\item{num}{(numeric vector) Numbers to display}

\item{pct}{(Boolean) Whether number is a proportion/percentage}

\item{digits}{(numeric) Number of decimal digits to display}
}
\description{
Display numbers in a pretty format
}
