% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{impute_covars}
\alias{impute_covars}
\title{Imputation procedure as described on PAP}
\usage{
impute_covars(d, covars, deact_impute = FALSE)
}
\arguments{
\item{d}{(data frame) Data frame that contains survey data}

\item{covars}{(character) Name(s) of covariates to be imputed}

\item{deact_impute}{(boolean) Whether to use mean imputation protocol outlined in
deactivation PAP}
}
\description{
The function imputes numerical covariates, including bins of 
factors, as the mean of non-missing values. For variables with 
>10% missing values, also create and include in LASSO covars a 
missingness binary (updates list of covariates to include 
missingness binary if needed).
}
