% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee_bounds.R
\name{leebounds}
\alias{leebounds}
\title{Estimating Lee (2009) treatment effect bounds}
\usage{
leebounds(leedata)
}
\arguments{
\item{leedata}{data frame containing three fields
\item{res}{leedata$treat: binary treatment indicator}
\item{res}{leedata$selection: selection=1 if the outcome is observed }
\item{res}{leedata$outcome:   outcome=selection*outcome}}
}
\value{
A list containing the estimate of lower bound and upper bound
}
\description{
\code{leebounds} basic Lee (2009) bounds on treatment effect without covariates. Sign of monotonicity is determined by trimming threshold.
}
