% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lin_model}
\alias{lin_model}
\title{Estimate treatment effects using a Lin (2013) regression model with lasso
covariate selection}
\usage{
lin_model(
  d,
  dv,
  dv_pre = NULL,
  lasso_covars,
  model_covars,
  trt,
  control,
  dv_label,
  weights_var,
  standardize = TRUE,
  proportion = FALSE,
  digits = 3,
  se_type = "HC2",
  impute = TRUE
)
}
\arguments{
\item{d}{(data frame) Data frame that contains the data}

\item{dv}{(character) Name of the DV}

\item{dv_pre}{(character) Name of the pre-treatment DV. \code{NULL} if not
available}

\item{lasso_covars}{(character vector) List of covariates to add to lasso
model for selection}

\item{model_covars}{(character vector) List of covariates to add to baseline
model, regardless of whether they were selected by the lasso or not.}

\item{trt}{(character) Name of treatment group}

\item{control}{(character) Name of control group}

\item{dv_label}{(character) Label for the DV, which can then be used to create
tables or graphs}

\item{weights_var}{(character) Name of the weights variable in the dataset}

\item{standardize}{(logical) If \code{TRUE}, the DV will be standardized
prior to fit the models, and the returning treatment effect estimate will be
a standardized coefficient.}

\item{proportion}{(logical) If \code{TRUE}, will multiply the DV by 100 prior
to fitting the model. Useful if the DV is a percentage; so that the output
is easier to read.}

\item{digits}{(numeric) Number of decimal digits to display in output.}

\item{se_type}{(character) Type of standard errors for treatment effect
estimates. This is inherited from estimatr. Valid types are: "HC0", "HC1",
"HC2", and "HC3". Default is "HC2"}

\item{impute}{(boolean) Whether to use mean imputation for missing values
of pretreatment covariates. Default is set to TRUE.}
}
\description{
This function follows a two-step procedure. First, it estimates a lasso model
where the DV is regressed on its pre-treatment measure (if available), as well
as a set of covariates measured prior to treatment. Then, it fits a Lin (2013)
model that includes the treatment indicator, the covariates selected by the
lasso model and additional covariates. The function returns both
weighted (PATE) and unweighted (SATE) treatment effects.
}
