% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{main_specification}
\alias{main_specification}
\title{Estimate treatment effects using main specification, given a subset of 
covariates and the type of treatment effects desired.}
\usage{
main_specification(
  type,
  d,
  dv,
  model_covars,
  selected_covars,
  dv_label = NULL,
  weights_var = NULL,
  iv = NULL,
  digits = 3,
  se_type = "HC2",
  leebounds = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{type}{(character) Model type to return.  Must be one of 'SATE',
'LATE', 'PLATE', or 'PATE'. The function returns outputs for a single type in 
each call}

\item{d}{(data frame) Data frame that contains the data}

\item{dv}{(character) Name of the DV}

\item{model_covars}{(character vector) List of covariates to add to baseline
model, regardless of whether they were selected by the lasso or not.}

\item{selected_covars}{(character vector) List of covariates to add to lasso
model for selection}

\item{dv_label}{(character) Label for the DV, which can then be used to create
tables or graphs. Default is same as \code{dv}.}

\item{weights_var}{(character) Name of the weights variable in the dataset.
Must be included if \code{type} is 'PATE', or 'PLATE'.}

\item{iv}{(character) Name in the dataset of the variable instrumented by the
treatment. Must be included if \code{type} is 'LATE' or 'PLATE'.}

\item{digits}{(numeric) Number of decimal digits to display in output.}

\item{se_type}{(character) Type of standard errors for treatment effect
estimates. This is inherited from estimatr. Valid types are: "HC0", "HC1",
"HC2", and "HC3". Default is "HC2"}

\item{leebounds}{(boolean) Whether to include Lee bounds for the estimates.}

\item{verbose}{(boolean) Whether to print messages with estimated effects.}

\item{dv_pre}{(character) Name of the pre-treatment DV. \code{NULL} if not
available. Default NULL.}
}
\description{
This function performs the main linear model specification, 
accepting different types of ATE: PATE, LATE, PLATE, SATE. 
The linear model includes the treatment indicator, the covariates selected
for the main specification (should be done by LASSO, see baseline_lasso) 
and additional covariates.
}
