% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_factor_bins}
\alias{make_factor_bins}
\title{Makes binary variables out of factors of factor variable(s).}
\usage{
make_factor_bins(d, covars)
}
\arguments{
\item{d}{(data frame) Data frame that contains survey data}

\item{covars}{(character) Name(s) of covariates (potential factors)}
}
\description{
The function turns factor covariates into binaries, excluding 
one of the binaries. The name of the new variables from a factor
`factor` with a value `Value One` is `factor_bin_value_one`. The 
function ouputs a dataframe with all binaries from factor variables
found on the input list of variables ($d). It also returns an updates
list of covariates that replaces the factor name with the names of 
the new factor binaries ($updated_covars)
}
