% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_treated_bin}
\alias{make_treated_bin}
\title{Make binary of whether observation is included in treated group}
\usage{
make_treated_bin(d, treatment_var = "treatment_group", trt)
}
\arguments{
\item{d}{(data frame) Data frame that contains survey data}

\item{treatment_var}{(character) Name of treatment group assignment variable.
Default 'treatment_group'}

\item{trt}{(character) Name of treatment group#'}
}
\description{
The function creates a binary variable, based on whether the 
treatment group assignment of the observation corresponds to the treated
group (trt).
}
