% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdata.R
\name{parse_map}
\alias{parse_map}
\title{Helper function to convert a map column to a list}
\usage{
parse_map(json)
}
\arguments{
\item{json}{(character) Vector with column that contains data uploaded from Facebook
in map format, and which needs to be converted to a list.}
}
\description{
Helper function to convert a map column to a list
}
\examples{
\dontrun{
d <- read_file(bucket='sandbox', filename='dim_op_fb_users.csv', has_header=TRUE)
d$vpv_count_category_map <- parse_map(d$vpv_count_category_map)
}
}
