% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passive.R
\name{parse_passive_data}
\alias{parse_passive_data}
\title{Parse raw passive monitoring data}
\usage{
parse_passive_data(min_date, max_date, verbose = TRUE)
}
\arguments{
\item{min_date}{(character) start date for data collection, in YYYY-MM-DD format}

\item{max_date}{(character) end date for data collection, in YYYY-MM-DD format}

\item{verbose}{(logical) If \code{TRUE} (default), show informative
messages indicating function progress}
}
\description{
Read raw passive monitoring files (for web visits and app usage)
and output aggregated daily tables, with different levels of
granularity, which will then be read by downstream functions.
}
\details{
Note that the domain-level dataset for web visits will only include domains
that were visited by 20 or more unique panelists. This k-anonymization
procedure was recommended by Privacy Engineering to reduce the risk of
re-identification of our panelists.

This function can only be run by Facebook researchers, since it requires
access to the visit-level passive monitoring data.
}
