% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{proportion_dvs}
\alias{proportion_dvs}
\title{Multiply variables by 100}
\usage{
proportion_dvs(d, dvs)
}
\arguments{
\item{d}{(data frame) Data frame that contains survey data}

\item{dvs}{(character) Name(s) of covariates to be transformed}
}
\description{
The function multiplies the variables with var_names on 
dvs vector by 100 and returns a dataframe with the transformed
variables
}
