% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{qa_table}
\alias{qa_table}
\title{Create a table with summary statistics for each variable in data.frame,
for debugging purposes}
\usage{
qa_table(data, to_exclude, output_file, verbose = TRUE)
}
\arguments{
\item{data}{(data frame) Data frame that contains relevant dataset}

\item{to_exclude}{(numeric vector) Columns in data frame that should be
excluded from table (e.g. character variables, DS, etc.)}

\item{output_file}{(character) Name of file where table will be stored (in
latex format)}

\item{verbose}{(logical) If \code{TRUE}, the function will output additional
information, such as the proportion of units with missing values.}
}
\description{
Create a table with summary statistics for each variable in data.frame,
for debugging purposes
}
