% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_file}
\alias{read_file}
\title{Download a specific file from a bucket and read it as a data.frame.}
\usage{
read_file(
  filename,
  has_header = FALSE,
  dlm = "guess",
  col_types = NULL,
  show_progress = FALSE
)
}
\arguments{
\item{filename}{(character) The name of the file to download}

\item{has_header}{Whether the file has a header or not}

\item{dlm}{(character) The delimiter to split the columns.  If ‘guess’ it uses the
end of the file (‘.tsv’,’.csv’) to infer the delimiter}

\item{col_types}{(character) Column types in file. \code{NULL} will impute
from the first 1000 rows in the input. See documentation for \code{read.csv}
for additional information.}

\item{show_progress}{(logical) Whether to show a progress bar}
}
\description{
Download a specific file from a bucket and read it as a data.frame.
}
