% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passive.R
\name{read_passive_data}
\alias{read_passive_data}
\title{Read aggregate passive monitoring files}
\usage{
read_passive_data(
  source,
  study_group = "all",
  study_platform = c("FB", "IG"),
  min_date = "2020-09-12",
  max_date = "2020-12-24",
  domain_dictionary,
  app_dictionary,
  verbose = TRUE
)
}
\arguments{
\item{source}{(character) 'web' or 'app', depending on the source of data
that needs to be aggregated}

\item{study_group}{(character vector) study groups for which respondent data
will be returned. Names of study groups must match those in database.}

\item{study_platform}{(character vector) study platforms for which respondent
data will be returned. Either \code{IG} or \code{FB}.}

\item{min_date}{(character) start date for data aggregation, in YYYY-MM-DD format}

\item{max_date}{(character) end date for data aggregation, in YYYY-MM-DD format}

\item{domain_dictionary}{(list) A name list where each key is a category
(e.g. 'misinfo') and the value is a vector with domains that fall in that
 category. Only if \code{granularity = 'domain'}.}

\item{app_dictionary}{(list) A name list where each key is a category
(e.g. 'misinfo') and the value is a vector with aop names that fall in that
 category. Only if \code{granularity = 'app'}}

\item{verbose}{(logical) If \code{TRUE} (default), show informative
messages indicating function progress}
}
\description{
This function reads the aggregate passive data files (for web visits and app usage)
and return a panelist-level dataset with counts of web visits or app opens
and time spent on apps, according to a set of filters.
}
\details{
Note that \code{study_group} for now will only work with platform intervention
groups (not deactivation).
}
