% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{standardize_dvs}
\alias{standardize_dvs}
\title{Standardization of set of variables}
\usage{
standardize_dvs(d, dvs, control)
}
\arguments{
\item{d}{(data frame) Data frame that contains survey data}

\item{dvs}{(character) Name(s) of covariates to be standardized}

\item{control}{(character) Name of control group}
}
\description{
The function standardizes the given set of variables, using the 
mean and standard deviation of the control group, returning
an updated dataframe with standardized covariates
}
