% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{theme_alethea}
\alias{theme_alethea}
\title{Project's ggplot2 theme}
\usage{
theme_alethea(
  base_size = 16,
  base_family = "Helvetica",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{(numeric) Base size}

\item{base_family}{(character) Base family}

\item{base_line_size}{(numeric) Base line size}

\item{base_rect_size}{(numeric) Base rect size}
}
\description{
Project's ggplot2 theme
}
