% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdata.R
\name{unmap}
\alias{unmap}
\title{Helper function to extract a value from a map column}
\usage{
unmap(map, key)
}
\arguments{
\item{map}{(list) Column within a data frame that corresponds to a map variable}

\item{key}{(character) Name of the key whose value needs to be extracted}
}
\description{
Helper function to extract a value from a map column
}
\examples{
\dontrun{
# Compute average VPVs of civic content
d <- read_file(bucket='sandbox', filename='dim_op_fb_users.csv', has_header=TRUE)
d$vpv_count_category_map <- parse_map(d$vpv_count_category_map)
mean(unmap(d$vpv_count_category_map, 'civic'), na.rm=TRUE)
}
}
