% Copyright (c) Meta Platforms, Inc. and affiliates.
% All rights reserved.

% This source code is licensed under the license found in the
% LICENSE file in the root directory of this source tree.

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize raw count at a given percentile threshold}
\usage{
winsorize(x, pct = 0.99)
}
\arguments{
\item{x}{(numeric) Number to winsorize}

\item{pct}{(numeric) Percentile at which to winsorize}
}
\description{
Winsorize raw count at a given percentile threshold
}
